requirements.txt
Django==4.2.11
djangorestframework==3.15.1
django-environ==0.11.2
django-extensions==3.2.3
ipython==8.18.1
psycopg2-binary==2.9.9
django-colorfield==0.11.0
Pillow==10.3.0